requirements.txt
flask==2.0.1
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
werkzeug==2.0.1
email-validator==1.1.3
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv==0.20.0 